Rename ComplianceService to MoodleService
