+++ api/src/test/java/org/pmiops/workbench/ras/RasLinkServiceTest.java
@@ -38,7 +38,6 @@ import org.pmiops.workbench.access.AccessTierService;
 import org.pmiops.workbench.access.UserAccessModuleMapperImpl;
 import org.pmiops.workbench.actionaudit.auditors.UserServiceAuditor;
 import org.pmiops.workbench.billing.FreeTierBillingService;
-import org.pmiops.workbench.compliance.ComplianceService;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.db.dao.AccessModuleDao;
 import org.pmiops.workbench.db.dao.UserAccessModuleDao;
@@ -56,6 +55,7 @@ import org.pmiops.workbench.identityverification.IdentityVerificationService;
 import org.pmiops.workbench.institution.InstitutionService;
 import org.pmiops.workbench.mail.MailService;
 import org.pmiops.workbench.model.Institution;
+import org.pmiops.workbench.moodle.MoodleService;
 import org.pmiops.workbench.test.FakeClock;
 import org.pmiops.workbench.test.FakeLongRandom;
 import org.pmiops.workbench.testconfig.UserServiceTestConfiguration;
@@ -158,7 +158,7 @@ public class RasLinkServiceTest {
   })
   @MockBean({
     AccessTierService.class,
-    ComplianceService.class,
+    MoodleService.class,
     DirectoryService.class,
     FireCloudService.class,
     FreeTierBillingService.class,
